+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.2.1 - 2015-XX-XX
 ### Change 
 - Rename calcalation classes. KISS =)
+